Update function example prompts
