+++ manifest.xml
@@ -6,7 +6,7 @@
   xmlns:ov="http://schemas.microsoft.com/office/taskpaneappversionoverrides"
   xsi:type="TaskPaneApp">
   <Id>f03a2d07-347f-4914-8d20-1be1a1812ccc</Id>
-  <Version>1.1</Version>
+  <Version>1.2.0</Version>
   <ProviderName>Octagon AI</ProviderName>
   <DefaultLocale>en-US</DefaultLocale>
   <DisplayName DefaultValue="Octagon AI - Market Intelligence"/>
