+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "excel-custom-functions",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/OfficeDev/Excel-Custom-Functions.git"
