+++ src/api/index.ts
@@ -1,3 +1,2 @@
-export { OCTAGON_AGENTS } from "./agents";
 export { OctagonApiService } from "./octagonApi";
 export { AgentType } from "./types";
