+++ src/taskpane/taskpane.html
@@ -70,39 +70,65 @@
     <main id="agents-view" class="main-content" style="display: none;">
         <div class="content-container">
 
-            <!-- Agent categories will be dynamically generated here -->
-            <div id="agent-categories-container"></div>
+            <div>
+                <div id="agent-card">
+                    <h4 class="agent-title">Octagon Agent</h4>
+                    <p class="agent-description">Public and Private market intelligence agent that optimally routes
+                        requests to appropriate specialized agents</p>
+                    <div class="agent-meta">
+                        <div>
+                            <strong>Function:</strong>
+                            <span class="formula-name">OCTAGON.AGENT(prompt, [format])</span>
+                            <div class="formula-argument">
+                                <code>prompt</code> is the question or prompt for the Octagon agent
+                            </div>
+                            <div class="formula-argument">
+                                <code>[format]</code> is optional with possible values of 'table' (default),
+                                'cell', or 'raw'.
+                                <br />
+                            </div>
+                            <div>
+                                Use 'table' for generating tabular data spanning multiple cells, 'cell' for a single
+                                cell value, and 'raw' for more verbose text-based content.
+                            </div>
+                        </div>
+                    </div>
+                    <div class="usage-examples-section">
+                        <h5 class="usage-heading">Usage Examples:</h5>
+                    </div>
+                    <!-- Agent categories will be dynamically generated here -->
+                </div>
 
-            <div class="back-button-container">
-                <button id="back-to-login-button" class="ms-Button ms-Button--default">
-                    <span class="ms-Button-icon">
-                        <i class="ms-Icon ms-Icon--ChevronLeft"></i>
-                    </span>
-                    <span class="ms-Button-label">Logout</span>
-                </button>
-            </div>
+                <div class="back-button-container">
+                    <button id="back-to-login-button" class="ms-Button ms-Button--default">
+                        <span class="ms-Button-icon">
+                            <i class="ms-Icon ms-Icon--ChevronLeft"></i>
+                        </span>
+                        <span class="ms-Button-label">Logout</span>
+                    </button>
+                </div>
 
-            <!-- Additional Resources Section -->
-            <div class="additional-resources">
-                <h3>Resources</h3>
-                <ul>
-                    <li>
-                        <i class="ms-Icon ms-Icon--ReadingMode"></i>
-                        <a href="https://docs.octagonagents.com/" target="_blank">Official Documentation</a>
-                    </li>
-                    <li>
-                        <i class="ms-Icon ms-Icon--TestBeaker"></i>
-                        <a href="https://app.octagonai.co/playground" target="_blank">API Playground</a>
-                    </li>
-                </ul>
+                <!-- Additional Resources Section -->
+                <div class="additional-resources">
+                    <h3>Resources</h3>
+                    <ul>
+                        <li>
+                            <i class="ms-Icon ms-Icon--ReadingMode"></i>
+                            <a href="https://docs.octagonagents.com/" target="_blank">Official Documentation</a>
+                        </li>
+                        <li>
+                            <i class="ms-Icon ms-Icon--TestBeaker"></i>
+                            <a href="https://app.octagonai.co/playground" target="_blank">API Playground</a>
+                        </li>
+                    </ul>
+                </div>
             </div>
-        </div>
 
     </main>
 
     <!-- Octagon Footer -->
     <div class="footer">
-        <div class="version">Version 1.1</div>
+        <div class="version">Version 1.2</div>
         <div class="footer-links">
             <a href="https://www.octagonai.co/terms-of-service" target="_blank">Terms of Service</a> |
             <a href="https://www.octagonai.co/privacy" target="_blank">Privacy</a> |
